+++ packages/core/src/utils/fs.ts
@@ -1,5 +1,7 @@
 import { readFileSync } from "fs";
 
+const STDIN_FILE_DESCRIPTOR = 0;
+
 export function readTextFromPathOrStdin(path: string): string {
-  return readFileSync(path === "-" ? "/dev/stdin" : path, "utf-8");
+  return readFileSync(path === "-" ? STDIN_FILE_DESCRIPTOR : path, "utf-8");
 }